Views code refactoring
